add TestingInfrastructureHelper class
